refactor(core): rename "communication" to "use_cases"
