+++ core/app/alarm/tests/test_on_connected_services.py
@@ -14,7 +14,7 @@ class OnConnectedServices(TestCase):
     def test_on_connect_send_data(self):
         handler = OnConnectedCamera(self.mqtt)
 
-        with patch('alarm.communication.out_alarm.notify_alarm_status_factory') as notify_alarm_status_factory:
+        with patch('alarm.use_cases.out_alarm.notify_alarm_status_factory') as notify_alarm_status_factory:
             notify = Mock()
             notify_alarm_status_factory.return_value = notify
