+++ core/app/alarm/use_cases/camera_motion.py
@@ -1,8 +1,8 @@
 import logging
 from alarm.mqtt.mqtt_data import InMotionCameraData 
 import alarm.business.in_motion as in_motion
-import alarm.communication.out_alarm as out_alarm
-from alarm.communication.play_sound import play_sound
+import alarm.use_cases.out_alarm as out_alarm
+from alarm.use_cases.play_sound import play_sound
 from alarm.models import AlarmStatus
 import alarm.notifications as alarm_notifications
