+++ core/app/alarm/use_cases/camera_picture.py
